auth-service/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
sqlalchemy==2.0.20
pymysql==1.1.0
cryptography==41.0.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1 